+++ Server/src/main/java/main/repository/SalesRepository.java
@@ -3,5 +3,5 @@ package main.repository;
 import main.entity.Sales;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SalesRepository extends CrudRepository<Sales, Integer> {
+public interface SalesRepository extends CrudRepository<Sales, Long> {
 }
