+++ Server/src/main/java/main/repository/Warehouse1Repository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface Warehouse1Repository extends CrudRepository<Warehouse1, Integer> {
+public interface Warehouse1Repository extends CrudRepository<Warehouse1, Long> {
     Optional<List<Warehouse1>> findAllById(Integer id);
 }
