+++ Server/src/main/java/main/service/GoodsService.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface GoodsService {
     List<Goods> listGoods();
-    Goods findGood(Integer id);
+    Goods findGood(Long id);
 }
